move getJavaConfig to utils.ts
